src/test/java/org/antonio/appmockito/ejemplos/services/ExamenServiceImplTest.java
package org.antonio.appmockito.ejemplos.services;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertNotNull;
import static org.junit.jupiter.api.Assertions.assertTrue;

import java.util.Arrays;
import java.util.List;
import java.util.Optional;

import org.antonio.appmockito.ejemplos.models.Examen;
import org.antonio.appmockito.ejemplos.repositories.ExamenRepository;
import org.antonio.appmockito.ejemplos.repositories.ExamenRepositoryImpl;
import org.junit.jupiter.api.Test;
import static org.mockito.Mockito.*;

public class ExamenServiceImplTest {
    @Test
    void testFindExamenPorNombre() {
        ExamenRepository repository = mock(ExamenRepository.class);
        ExamenService service = new ExamenServiceImpl(repository);
        List<Examen> datos = Arrays.asList(new Examen(5L, "Matemáticas"), new Examen(6L, "Lenguaje"),
                new Examen(7L, "Historia"));

        when(repository.findAll()).thenReturn(datos);
        Optional<Examen> examen = service.findExamenPorNombre("Matemáticas");

        assertTrue(examen.isPresent());
        assertEquals(5L, examen.orElseThrow(null).getId());
        assertEquals("Matemáticas", examen.get().getNombre());
    }
}
